# tp-final/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/tenants/{tenantSlug}/clientes/codigos": {
      "post": {
        "operationId": "ClientesController_pedirCodigo",
        "parameters": [
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedirCodigoDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodigoPedidoDto"
                }
              }
            }
          },
          "400": {
            "description": "El email no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "429": {
            "description": "Demasiados codigos para ese email, o demasiadas peticiones.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Manda un codigo de ingreso de seis digitos al email de la clienta.",
        "tags": [
          "Clientes"
        ]
      }
    },
    "/api/v1/tenants/{tenantSlug}/clientes/sesiones": {
      "post": {
        "operationId": "ClientesController_ingresar",
        "parameters": [
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrearSesionClienteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SesionClienteDto"
                }
              }
            }
          },
          "401": {
            "description": "Codigo invalido, vencido, ya usado o con los intentos agotados.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "summary": "Entra con el codigo. Si es la primera vez, crea la cuenta.",
        "tags": [
          "Clientes"
        ]
      }
    },
    "/api/v1/tenants/{tenantSlug}/clientes/me": {
      "get": {
        "operationId": "ClientesController_findMe",
        "parameters": [
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro, o no es de una clienta.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "El perfil de la clienta de la sesion.",
        "tags": [
          "Clientes"
        ]
      },
      "patch": {
        "operationId": "ClientesController_updateMe",
        "parameters": [
          {
            "name": "tenantSlug",
            "required": true,
            "in": "path",
            "description": "Identificador del centro en la URL, por ejemplo `lo-de-lili`.",
            "schema": {
              "example": "lo-de-lili",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClienteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClienteDto"
                }
              }
            }
          },
          "400": {
            "description": "Datos invalidos.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "401": {
            "description": "Falta el token o no es valido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "403": {
            "description": "El token es de otro centro, o no es de una clienta.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "El centro no existe o esta inactivo.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Edita nombre y telefono. El email no se cambia: es la identidad de la cuenta.",
        "tags": [
          "Clientes"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de turnos",
    "description": "SaaS multi-tenant de turnos para centros de estetica. Cada centro es un tenant y viaja en el path. Los montos son enteros en centavos. Las horas son hora de pared del centro, en formato HH:mm, y las fechas YYYY-MM-DD.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Identificador de maquina del error. Es el que conviene usar para ramificar.",
            "enum": [
              "validation_error",
              "invalid_cursor",
              "invalid_credentials",
              "unauthenticated",
              "invalid_token",
              "wrong_tenant",
              "tenant_not_found",
              "servicio_not_found",
              "reserva_not_found",
              "servicio_name_taken",
              "ventanas_superpuestas",
              "slot_full",
              "outside_business_hours",
              "past_date",
              "too_far_ahead",
              "invalid_transition",
              "high_contention",
              "forbidden_role",
              "not_found",
              "payload_too_large",
              "too_many_requests",
              "internal_error",
              "http_error",
              "invalid_code",
              "too_many_codes",
              "cliente_not_found",
              "online_payment_unavailable",
              "payment_provider_unavailable",
              "reschedule_not_allowed",
              "too_early_for_no_show",
              "invalid_state",
              "mp_account_change_blocked",
              "mercadopago_not_configured",
              "monthly_limit_reached",
              "plan_limit_reached",
              "mercadopago_not_connected",
              "invalid_signature",
              "slug_taken"
            ],
            "type": "string",
            "example": "slot_full"
          },
          "message": {
            "type": "string",
            "description": "Texto en espanol, listo para mostrarle a una persona.",
            "example": "Ese horario ya no tiene cupo. Elegi otro."
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "PedirCodigoDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "El email de la clienta. Si no tiene cuenta en el centro, se crea al entrar.",
            "format": "email",
            "example": "ana@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "CodigoPedidoDto": {
        "type": "object",
        "properties": {
          "expiraEnMinutos": {
            "type": "number",
            "description": "Cuanto tarda en vencer el codigo que se acaba de mandar.",
            "example": 10
          }
        },
        "required": [
          "expiraEnMinutos"
        ]
      },
      "CrearSesionClienteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "El email de la clienta. Si no tiene cuenta en el centro, se crea al entrar.",
            "format": "email",
            "example": "ana@example.com"
          },
          "codigo": {
            "type": "string",
            "description": "Los seis digitos que llegaron por mail.",
            "pattern": "^\\d{6}$",
            "example": "482913"
          }
        },
        "required": [
          "email",
          "codigo"
        ]
      },
      "ClienteDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "La identidad de la clienta en el centro. No se cambia: es a donde llega el codigo.",
            "example": "ana@example.com"
          },
          "id": {
            "type": "string"
          },
          "nombre": {
            "type": "string",
            "nullable": true,
            "description": "null hasta que lo complete, reservando o desde el perfil."
          },
          "telefono": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "email",
          "id",
          "nombre",
          "telefono"
        ]
      },
      "SesionClienteDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT para el header `Authorization: Bearer <token>`. Dura 30 dias: una clienta entra poco,\ny cada ingreso le cuesta ir al mail a buscar el codigo.",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "cliente": {
            "$ref": "#/components/schemas/ClienteDto"
          }
        },
        "required": [
          "accessToken",
          "cliente"
        ]
      },
      "UpdateClienteDto": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "telefono": {
            "type": "string",
            "minLength": 6,
            "maxLength": 30
          }
        }
      },
...
    }
  }
}